@@ -12,53 +12,52 @@ use Unleashed\Helpers\UseStatements;
 
 final class FullyQualifiedGlobalFunctionsSniff implements Sniff
 {
-    /** @var bool */
-    public $onlyOptimizedFunctions = false;
+    public bool $onlyOptimizedFunctions = false;
 
     /** @var array<string, bool> */
-    private $optimizedFunctions = [
+    private array $optimizedFunctions = [
         // @see https://github.com/php/php-src/blob/PHP-7.4/Zend/zend_compile.c "zend_try_compile_special_func"
-        'array_key_exists'     => true,
-        'array_slice'          => true,
-        'assert'               => true,
-        'boolval'              => true,
-        'call_user_func'       => true,
+        'array_key_exists' => true,
+        'array_slice' => true,
+        'assert' => true,
+        'boolval' => true,
+        'call_user_func' => true,
         'call_user_func_array' => true,
-        'chr'                  => true,
-        'count'                => true,
-        'defined'              => true,
-        'doubleval'            => true,
-        'floatval'             => true,
-        'func_get_args'        => true,
-        'func_num_args'        => true,
-        'get_called_class'     => true,
-        'get_class'            => true,
-        'gettype'              => true,
-        'in_array'             => true,
-        'intval'               => true,
-        'is_array'             => true,
-        'is_bool'              => true,
-        'is_double'            => true,
-        'is_float'             => true,
-        'is_int'               => true,
-        'is_integer'           => true,
-        'is_long'              => true,
-        'is_null'              => true,
-        'is_object'            => true,
-        'is_real'              => true,
-        'is_resource'          => true,
-        'is_string'            => true,
-        'ord'                  => true,
-        'sizeof'               => true,
-        'strlen'               => true,
-        'strval'               => true,
+        'chr' => true,
+        'count' => true,
+        'defined' => true,
+        'doubleval' => true,
+        'floatval' => true,
+        'func_get_args' => true,
+        'func_num_args' => true,
+        'get_called_class' => true,
+        'get_class' => true,
+        'gettype' => true,
+        'in_array' => true,
+        'intval' => true,
+        'is_array' => true,
+        'is_bool' => true,
+        'is_double' => true,
+        'is_float' => true,
+        'is_int' => true,
+        'is_integer' => true,
+        'is_long' => true,
+        'is_null' => true,
+        'is_object' => true,
+        'is_real' => true,
+        'is_resource' => true,
+        'is_string' => true,
+        'ord' => true,
+        'sizeof' => true,
+        'strlen' => true,
+        'strval' => true,
         // @see https://github.com/php/php-src/blob/php-7.2.6/ext/opcache/Optimizer/pass1_5.c
-        'constant'             => true,
-        'define'               => true,
-        'dirname'              => true,
-        'extension_loaded'     => true,
-        'function_exists'      => true,
-        'is_callable'          => true,
+        'constant' => true,
+        'define' => true,
+        'dirname' => true,
+        'extension_loaded' => true,
+        'function_exists' => true,
+        'is_callable' => true,
     ];
 
     /**
@@ -100,20 +99,20 @@ final class FullyQualifiedGlobalFunctionsSniff implements Sniff
 
         $whitelist = UseStatements::getAliasesAndNonGlobalFunctionsDefinedInUseStatements($phpcsFile);
 
-        $tokens    = $phpcsFile->getTokens();
-        $ignore    = [
-            T_DOUBLE_COLON    => true,
+        $tokens = $phpcsFile->getTokens();
+        $ignore = [
+            T_DOUBLE_COLON => true,
             T_OBJECT_OPERATOR => true,
-            T_FUNCTION        => true,
-            T_CONST           => true,
-            T_PUBLIC          => true,
-            T_PRIVATE         => true,
-            T_PROTECTED       => true,
-            T_AS              => true,
-            T_NEW             => true,
-            T_INSTEADOF       => true,
-            T_NS_SEPARATOR    => true,
-            T_IMPLEMENTS      => true,
+            T_FUNCTION => true,
+            T_CONST => true,
+            T_PUBLIC => true,
+            T_PRIVATE => true,
+            T_PROTECTED => true,
+            T_AS => true,
+            T_NEW => true,
+            T_INSTEADOF => true,
+            T_NS_SEPARATOR => true,
+            T_IMPLEMENTS => true,
         ];
         $prevToken = $phpcsFile->findPrevious([T_WHITESPACE, T_COMMENT], $stackPtr - 1, null, true);
 
@@ -138,7 +137,7 @@ final class FullyQualifiedGlobalFunctionsSniff implements Sniff
             return;
         }
 
-        $function           = \strtolower($tokens[$stackPtr]['content']);
+        $function = \strtolower($tokens[$stackPtr]['content']);
         $functionNormalized = \strtolower($function);
 
         // Is it a whitelisted alias?
@@ -152,7 +151,7 @@ final class FullyQualifiedGlobalFunctionsSniff implements Sniff
         }
 
         $error = \sprintf('Function %1$s() should be referenced via a fully qualified name, e.g.: \%1$s()', $function);
-        $fix   = $phpcsFile->addFixableError($error, $stackPtr, 'NotFullyQualified');
+        $fix = $phpcsFile->addFixableError($error, $stackPtr, 'NotFullyQualified');
 
         if ($fix === true) {
             $this->applyFix($phpcsFile, $stackPtr, $function);
