@@ -14,14 +14,11 @@ use PHP_CodeSniffer\Files\File;
 final class SniffLocalCache
 {
     /** @var array<int, array<string, mixed>> */
-    private static $cache = [];
+    private static array $cache = [];
 
-    /**
-     * @return mixed
-     */
-    public static function getAndSetIfNotCached(File $phpcsFile, string $key, \Closure $lazyValue)
+    public static function getAndSetIfNotCached(File $phpcsFile, string $key, \Closure $lazyValue): mixed
     {
-        $fixerLoops  = $phpcsFile->fixer !== null ? $phpcsFile->fixer->loops : 0;
+        $fixerLoops = $phpcsFile->fixer?->loops ?? 0;
         $internalKey = \sprintf('%s-%s', $phpcsFile->getFilename(), $key);
 
         self::setIfNotCached($fixerLoops, $internalKey, $lazyValue);
