@@ -14,7 +14,7 @@ use SlevomatCodingStandard\Helpers\TokenHelper;
 final class DescriptionRequiredSniff implements Sniff
 {
     public const CODE_MISSING_DESCRIPTION = 'MissingDescription';
-    public const CODE_EMPTY_DESCRIPTION   = 'EmptyDescription';
+    public const CODE_EMPTY_DESCRIPTION = 'EmptyDescription';
 
     /**
      * Returns an array of tokens this test wants to listen for.
@@ -46,7 +46,7 @@ final class DescriptionRequiredSniff implements Sniff
 
         // Does a `getDescription()` method exist?
         $classEndPtr = self::findApproximateClassEndPointer($phpcsFile, $parentClassPtr);
-        $methodPtr   = self::findMethodInClass($phpcsFile, 'getDescription', $parentClassPtr, $classEndPtr);
+        $methodPtr = self::findMethodInClass($phpcsFile, 'getDescription', $parentClassPtr, $classEndPtr);
         if ($methodPtr === null) {
             // Nope - method is missing
             $fix = $phpcsFile->addFixableError(
@@ -80,7 +80,7 @@ final class DescriptionRequiredSniff implements Sniff
         );
     }
 
-    private static function findMethodInClass(File $phpcsFile, string $methodName, int $startPtr, int $endPtr): ?int
+    private static function findMethodInClass(File $phpcsFile, string $methodName, int $startPtr, int $endPtr): int|null
     {
         do {
             $nextFunctionPointer = TokenHelper::findNext($phpcsFile, T_FUNCTION, $startPtr + 1);
@@ -110,13 +110,13 @@ final class DescriptionRequiredSniff implements Sniff
         return null;
     }
 
-    private static function findApproximateClassEndPointer(File $phpcsFile, int $ptrWithinClass): ?int
+    private static function findApproximateClassEndPointer(File $phpcsFile, int $ptrWithinClass): int|null
     {
         $classPtrs = \array_keys(ClassHelper::getAllNames($phpcsFile));
 
         while (\current($classPtrs) !== false) {
             $start = \current($classPtrs);
-            $end   = \next($classPtrs);
+            $end = \next($classPtrs);
 
             if ($start <= $ptrWithinClass && $end >= $ptrWithinClass) {
                 return $end;
